Scan entire timeline
Due to the integer math for determining the number of timeline pages to
check, the last page (with the oldest) tweets would normally be missed.

For example: 743 tweet. Max 250 per page.

743 / 250 = 2

What we want, though is 3.

This commit changes MAX_TWEETS_PER_PAGE to 250.0 from 250. Now,

743 / 250.0 = 2.972

We then call `ceil` on it to round up to the next highest integer, 3.0
and correctly check all pages.